+++ src/waveline/_common.py
@@ -1,12 +1,11 @@
 import re
 from typing import List
 
-
 # key = value pattern for ae/tr data
 # fast(est) and simple, accept spaces around "="
 # _KV_PATTERN = re.compile(br"(\w+)\s*=\s*(\S+)")
 # accept words as keys w/o values; this seems next faster (incl. \S)?!
-KV_PATTERN = re.compile(br"([^\s=]+)(?:\s*=\s*(\S+))?")
+KV_PATTERN = re.compile(rb"([^\s=]+)(?:\s*=\s*(\S+))?")
 
 
 def as_int(string, default: int = 0):