+++ tests/unit/test_conditionwave.py
@@ -2,10 +2,16 @@ import asyncio
 from typing import NamedTuple
 from unittest.mock import call, patch
 
+import numpy as np
 import pytest
 from asynctest import CoroutineMock, Mock  # alternative to unittest.mock.AsyncMock in Python >=3.8
+from numpy.testing import assert_allclose
 
 from waveline import ConditionWave
+
+CLOCK = 10e6
+ADC_TO_VOLTS = [1.5625e-06, 0.00015625]
+ADC_TO_EU = [factor ** 2 * 1e14 / CLOCK for factor in ADC_TO_VOLTS]
 
 # all test coroutines will be treated as marked
 pytestmark = pytest.mark.asyncio
@@ -20,18 +26,27 @@ class MockedObjects(NamedTuple):
 
 
 @pytest.fixture(autouse=True, name="mock_objects")
-async def mock_asyncio_connection():
+async def mock_conditionwave_with_asyncio_connection():
     with patch("asyncio.open_connection", new=CoroutineMock()) as mock_create_connection:
         mock_reader = Mock(spec=asyncio.StreamReader)
         mock_writer = Mock(spec=asyncio.StreamWriter)
         mock_create_connection.return_value = (mock_reader, mock_writer)
 
-        async with ConditionWave("192.168.0.100") as cw:
-            yield MockedObjects(
-                cw=cw,
-                reader=mock_reader,
-                writer=mock_writer,
-            )
+        # ConditionWave.connect will call _check_firmware_version and _get_adc_to_volts
+        with patch.object(
+            ConditionWave, "_check_firmware_version", new=CoroutineMock()
+        ) as method_firmware_version, patch.object(
+            ConditionWave, "_get_adc_to_volts", new=CoroutineMock()
+        ) as method_adc_to_volts:
+            method_firmware_version.return_value = None
+            method_adc_to_volts.return_value = ADC_TO_VOLTS
+
+            async with ConditionWave("192.168.0.100") as cw:
+                yield MockedObjects(
+                    cw=cw,
+                    reader=mock_reader,
+                    writer=mock_writer,
+                )
 
 
 async def test_reconnect(mock_objects):
@@ -131,6 +146,19 @@ async def test_get_setup(mock_objects):
     assert setup.tr_decimation == 2
     assert setup.tr_pretrigger_samples == 100
     assert setup.tr_postduration_samples == 50
+
+
+@pytest.mark.parametrize(
+    "channel, enabled, command",
+    (
+        (0, True, b"set_acq enabled 1 @0\n"),
+        (1, False, b"set_acq enabled 0 @1\n"),
+    ),
+)
+async def test_set_channel(mock_objects, channel, enabled, command):
+    cw, _, writer = mock_objects
+    await cw.set_channel(channel, enabled)
+    writer.write.assert_called_with(command)
 
 
 @pytest.mark.parametrize(
@@ -254,6 +282,12 @@ async def test_set_filter(mock_objects, channel, highpass, lowpass, order, comma
     writer.write.assert_called_with(command)
 
 
+@pytest.mark.parametrize("channel", (-1, 3))
+async def test_set_filter_invalid_channel(mock_objects, channel):
+    with pytest.raises(ValueError):
+        await mock_objects.cw.set_filter(channel, 50e3, 300e3, 4)
+
+
 @pytest.mark.parametrize(
     "channel, samples, command",
     (
@@ -293,15 +327,113 @@ async def test_set_threshold(mock_objects, channel, threshold, command):
     writer.write.assert_called_with(command)
 
 
-@pytest.mark.parametrize("channel", (-1, 3))
-async def test_set_filter_invalid_channel(mock_objects, channel):
-    with pytest.raises(ValueError):
-        await mock_objects.cw.set_filter(channel, 50e3, 300e3, 4)
-
-
 async def test_start_stop_acquisition(mock_objects):
     cw, _, writer = mock_objects
     await cw.start_acquisition()
     writer.write.assert_called_with(b"start_acq\n")
     await cw.stop_acquisition()
     writer.write.assert_called_with(b"stop_acq\n")
+
+
+@pytest.mark.parametrize(
+    "channel, interval, count, cycles, command",
+    (
+        (0, 1, 4, 1, b"start_pulsing 1 4 1 @0\n"),
+        (1, 0.1, 0, 0, b"start_pulsing 0.1 0 0 @1\n"),
+    ),
+)
+async def test_start_pulsing(mock_objects, channel, interval, count, cycles, command):
+    cw, _, writer = mock_objects
+    await cw.start_pulsing(channel, interval, count, cycles)
+    writer.write.assert_called_with(command)
+
+
+async def test_stop_pulsing(mock_objects):
+    cw, _, writer = mock_objects
+    await cw.stop_pulsing()
+    writer.write.assert_called_with(b"stop_pulsing\n")
+
+
+async def test_get_ae_data(mock_objects):
+    cw, reader, writer = mock_objects
+    reader.readline.side_effect = [
+        b"S Ch=1 T=20000000 A=0 R=0 D=10000000 C=0 E=38788614 TRAI=0 flags=0\n",
+        b"H Ch=2 T=43686000 A=31004 R=496 D=703 C=4 E=74860056830 TRAI=1 flags=0\n",
+        b"\n",
+        wait_forever(),
+    ]
+
+    ae_data = await cw.get_ae_data()
+    writer.write.assert_called_with(b"get_ae_data\n")
+    reader.readline.assert_awaited()
+
+    assert len(ae_data) == 2
+
+    adc_to_volts = ADC_TO_VOLTS[0]
+    adc_to_eu = ADC_TO_EU[0]
+
+    # status record
+    s = ae_data[0]
+    assert s.type_ == "S"
+    assert s.channel == 1
+    assert s.time == pytest.approx(20000000 / CLOCK)
+    assert s.amplitude == 0
+    assert s.rise_time == 0
+    assert s.duration == pytest.approx(10000000 / CLOCK)
+    assert s.counts == 0
+    assert s.energy == pytest.approx(38788614 * adc_to_eu)
+    assert s.trai == 0
+    assert s.flags == 0
+
+    # hit record
+    h = ae_data[1]
+    assert h.type_ == "H"
+    assert h.channel == 2
+    assert h.time == pytest.approx(43686000 / CLOCK)
+    assert h.amplitude == pytest.approx(31004 * adc_to_volts)
+    assert h.rise_time == pytest.approx(496 / CLOCK)
+    assert h.duration == pytest.approx(703 / CLOCK)
+    assert h.counts == 4
+    assert h.energy == pytest.approx(74860056830 * adc_to_eu)
+    assert h.trai == 1
+    assert h.flags == 0
+
+
+@pytest.mark.parametrize("raw", (False, True))
+async def test_get_tr_data(mock_objects, raw):
+    cw, reader, writer = mock_objects
+    reader.readline.side_effect = [
+        b"Ch=1 TRAI=1 T=43686000 NS=13\n",
+        b"Ch=2 TRAI=2 T=43686983 NS=27\n",
+        b"\n",
+        wait_forever(),
+    ]
+    data = [np.arange(samples, dtype=np.int16) for samples in (13, 27)]
+    binary_data = [arr.tobytes() for arr in data]
+
+    reader.readexactly.side_effect = binary_data
+
+    tr_data = await cw.get_tr_data(raw=raw)
+    writer.write.assert_called_with(b"get_tr_data\n")
+
+    adc_to_volts = ADC_TO_VOLTS[0]
+
+    assert tr_data[0].channel == 1
+    assert tr_data[0].trai == 1
+    assert tr_data[0].time == pytest.approx(43686000 / CLOCK)
+    assert tr_data[0].samples == 13
+    assert tr_data[0].raw == raw
+    if raw:
+        assert_allclose(tr_data[0].data, data[0])
+    else:
+        assert_allclose(tr_data[0].data, data[0] * adc_to_volts)
+
+    assert tr_data[1].channel == 2
+    assert tr_data[1].trai == 2
+    assert tr_data[1].time == pytest.approx(43686983 / CLOCK)
+    assert tr_data[1].samples == 27
+    assert tr_data[1].raw == raw
+    if raw:
+        assert_allclose(tr_data[1].data, data[1])
+    else:
+        assert_allclose(tr_data[1].data, data[1] * adc_to_volts)